+++ netharn/examples/classification.py
@@ -618,7 +618,7 @@ class ClfHarn(nh.FitHarn):
             >>> harn._demo_epoch('vali', max_iter=10)
             >>> harn.on_epoch()
         """
-        from netharn.metrics import clf_report
+        from kwcoco.metrics import clf_report
         dset = harn.datasets[harn.current_tag]
 
         probs = np.vstack(harn._accum_confusion_vectors['probs'])
@@ -628,7 +628,7 @@ class ClfHarn(nh.FitHarn):
         # _pred = probs.argmax(axis=1)
         # assert np.all(_pred == y_pred)
 
-        # from netharn.metrics import confusion_vectors
+        # from kwcoco.metrics import confusion_vectors
         # cfsn_vecs = confusion_vectors.ConfusionVectors.from_arrays(
         #     true=y_true, pred=y_pred, probs=probs, classes=dset.classes)
         # report = cfsn_vecs.classification_report()