+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Changed
 * Moved imgaug from runtime dependencies to problematic (because of opencv)
+* Removed netharn.metrics, use `kwcoco.metrics` instead
 
 ## Version 0.5.16 - Released 2021-06-16